refactor(ai): scaffold aiProvider modular layout and shim
