+++ backend/src/aiProvider/adapters/anthropic.js
@@ -1,0 +1,11 @@
+export async function generate() {
+  throw new Error("Adapter wiring pending: anthropic");
+}
+
+export async function stream() {
+  return null;
+}
+
+export async function generateVision() {
+  throw new Error("Adapter wiring pending: anthropic");
+}
