+++ backend/src/aiProvider/adapters/openai.js
@@ -1,0 +1,11 @@
+export async function generate() {
+  throw new Error("Adapter wiring pending: openai");
+}
+
+export async function stream() {
+  return null;
+}
+
+export async function generateVision() {
+  throw new Error("Adapter wiring pending: openai");
+}
