+++ backend/src/aiProvider/registry.js
@@ -1,0 +1,12 @@
+export {
+  setRuntimeKey,
+  setRuntimeOllama,
+  setActiveProvider,
+  getProvider,
+  hasProvider,
+  isLocalProvider,
+  isProviderDegraded,
+  getProviderMeta,
+  getProviderName,
+  loadKeysFromDatabase,
+} from "./index.js";
