+++ backend/tests/run-tests.js
@@ -72,6 +72,7 @@ const files = [
   "tests/ai-fallback.test.js",
   "tests/openrouter-provider.test.js",
   "tests/openai-compat-provider.test.js",
+  "tests/aiProvider-adapter-contract.test.js",
   "tests/compat-config-cache.test.js",
   "tests/api-versioning.test.js",
   "tests/robots-sitemap.test.js",
